Entities are ready to be persisted to DB individually. Correct order …

+++ src/main/java/app/entities/Movie.java
@@ -25,7 +25,7 @@ private String releaseDate;
     @Column(nullable = false)
 private String description;
     @Column(nullable = false)
-    private Integer populatity;
+    private double populatity;
 
  /*   @ManyToMany
     @ToString.Exclude
